[Frontend]: Renamed 'analytics --> report'.
