+++ React-frontend/src/App.js
@@ -21,9 +21,8 @@ import CaseDirsPage from './pages/Management/CaseDirsPage';
 import CaseFilePage from './pages/Management/CaseFilePage';
 import CompletedtaskPage from './pages/Management/CompletedTaskPage';
 import TodoTaskPage from './pages/Management/TodoTaskPage';
-import AnalyticsMain from './pages/Management/Analytics/AnalyticsMain';
-import DataVisualizer from './pages/Management/Analytics/DataVisualizer';
-import Report from './components/Management/Report';
+import ReportMain from './pages/Management/Report/ReportMain';
+import DataVisualizer from './pages/Management/Report/DataVisualizer';
 
 
 
@@ -84,14 +83,13 @@ function App() {
           <PrivateRoute path='/task/create' exact component={CreateTask} />
           <PrivateRoute path='/task/list' exact component={ShowTasks} />
           <PrivateRoute path='/case-tree' exact component={CaseTreePage} />
-          <PrivateRoute path='/analytics' exact component={AnalyticsMain} />
-          <PrivateRoute path='/analytics/:caseName' component={DataVisualizer} />
+          <PrivateRoute path='/report' exact component={ReportMain} />
+          <PrivateRoute path='/report/:caseName' component={DataVisualizer} />
           <PrivateRoute path='/file-explorer' exact component={FileExplorerPage} />
           <PrivateRoute path='/file-explorer/:caseName/:dirName' component={CaseFilePage} />
           <PrivateRoute path='/file-explorer/:caseName' component={CaseDirsPage} />
           <PrivateRoute path='/task/finished' exact component={CompletedtaskPage} />
           <PrivateRoute path='/task/todo' exact component={TodoTaskPage} />
-          <PrivateRoute path='/report' exact component={Report} />
           <PublicRoute restricted={false} component={NotFound} />
         </Switch>
       </div>
