+++ React-frontend/src/components/Management/Report.js
@@ -1,62 +1,98 @@
-import React, { useState } from "react";
-import { makeStyles } from "@material-ui/core/styles";
-import { useDispatch, useSelector } from "react-redux";
-import { loadAnalyticsCommonWord } from "../../store/actions/management";
-import { withStyles } from "@material-ui/core/styles";
-import { Container } from "@material-ui/core";
+/*
+    Component to render all the cases to analyse.
+*/
+
+import React, { useEffect } from 'react';
+import { makeStyles } from '@material-ui/core/styles';
+import { useSelector, useDispatch } from 'react-redux';
+import { loadCases } from '../../store/actions/case';
+// import { useHistory } from 'react-router-dom';
+import CaseFolderButton from '../Utils/CaseFolderButton';
+import {
+    Container,
+    Typography,
+    Box,
+    Divider,
+    Button
+} from '@material-ui/core';
+
 
 // custom styles
-const useStyles = makeStyles((theme) => ({
-  root: {
-    width: "80vw",
-    overflowX: "auto",
-    minWidth: "40vw",
-    maxWidth: "90vw",
-    marginTop: "10vh",
-    height: "82.5vh",
-    padding: `${theme.spacing(1)}px ${theme.spacing(2)}px`,
-    display: "flex",
-    flexDirection: "column",
-    justifyContent: "flex-start",
-    alignItems: "flex-start",
-  },
-  input: {
-    marginBottom: theme.spacing(4),
-  },
-  inputs: {
-    height: theme.spacing(6),
-    width: theme.spacing(80),
-    marginTop: theme.spacing(1),
-    marginBottom: theme.spacing(2),
-  },
-  submit: {
-    color: theme.palette.text.main,
-    backgroundColor: theme.palette.primary.extraLight,
-    margin: theme.spacing(1.5, 0),
-    width: theme.spacing(80),
-    "&:hover": {
-      backgroundColor: theme.palette.primary.light,
+const useStyle = makeStyles((theme) => ({
+    root: {
+        width: '80vw',
+        overflowX: 'auto',
+        marginTop: '10vh',
+        height: '82.5vh',
+        padding: `${theme.spacing(1)}px ${theme.spacing(2)}px`,
+        display: 'flex',
+        flexDirection: 'column',
+        justifyContent: 'flex-start',
+        alignItems: 'flex-start'
     },
-  },
-  paper: {
-    marginTop: theme.spacing(2),
-    display: "flex",
-    overflowX: "auto",
-    minWidth: "35vw",
-    maxWidth: "85vw",
-  },
-  table: {
-    minWidth: 650,
-  },
-}));
+    title: {
+        fontSize: '1rem',
+        fontWeight: 'bolder',
+        '&:focus': {
+            outline: 'none'
+        }
+    },
+    caseList: {
+        marginTop: theme.spacing(2),
+        display: 'flex',
+        flexWrap: 'wrap',
+        justifyContent: 'center'
+    }
+}))
 
 function Report() {
-    const classes = useStyles();
+
+    // Dispatcher
+    const dispatch = useDispatch()
+
+    // useEffect
+    useEffect(() => {
+        dispatch(loadCases())
+    }, [dispatch])
+
+    // invoke custom styles
+    const classes = useStyle()
+
+    // get case reducer
+    const caseReducer = useSelector(state => state.case)
+
+    // history
+    // const history = useHistory()
+
+    // Main return statement
     return (
         <Container component="main" className={classes.root}>
-        Report Page
+
+            {/* Header */}
+            <Typography component="h1" variant="h5">
+                <Button className={classes.title}>
+                    Case Analyser
+                </Button>
+            </Typography>
+
+            {/* Divider */}
+            <Divider style={{width: '100%', marginTop: '1em'}} />
+
+            {/* Rendering all cases from caseReducer */}
+            <Box component="div" className={classes.caseList}>
+                {
+                    (caseReducer.cases && caseReducer.cases.length > 0) ?
+                        caseReducer.cases.map((caseItem) => {
+                            return (<CaseFolderButton key={caseItem.case_name} dirName={caseItem.case_name} parentDir='report' isAnalyser={true} />)
+                        })
+                    : (
+                        <span> Cases Not Found.</span>
+                    )
+                }
+            </Box>
+
         </Container>
     )
 }
 
-export default Report;
+export default Report
